jwt설정 변경
