+++ module-common/src/main/resources/application.properties
@@ -8,10 +8,6 @@ spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.properties.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
 
-token.secret.key=cgxz5nCh3+4g1p5pLlCeDft6gT4zqPdc9Vr2vym8NE0=
-token.access-token.plus-hour=1
-token.refresh-token.plus-hour=12
-
 springdoc.swagger-ui.path=/swagger-ui.html
 springdoc.api-docs.path=/v3/api-docs
 springdoc.packages-to-scan=com.welcommu.moduleapi
